JUnit created

+++ ec.edu.espol.workshops.second/src/main/java/ec/edu/espol/workshops/second/CarInsurance.java
@@ -2,7 +2,7 @@ package ec.edu.espol.workshops.second;
 
 public class CarInsurance {
 
-  private Double base = 500.0;
+  private Integer base = 500;
   private Customer customer;
 
   public void setCustomer(Customer customer) {
@@ -22,9 +22,9 @@ public class CarInsurance {
       this.customer.getAge() > 45 && this.customer.getAge() < 65;
     boolean married = this.customer.isMarried();
     if (male && young && !married) {
-      base += 1500.0;
+      base += 1500;
     } else if (!male || married) {
-      base -= 200.0;
+      base -= 200;
     } else if (notTooOld) {
       base -= 100;
     }
@@ -38,7 +38,7 @@ public class CarInsurance {
     return this.customer;
   }
 
-  public Double getBase() {
+  public Integer getBase() {
     return this.base;
   }
 }
